% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_get_domain.R
\name{bps_get_domain}
\alias{bps_get_domain}
\title{Get list of domain}
\usage{
bps_get_domain(type = "all", ..., token)
}
\arguments{
\item{type}{Type to show domain}

\item{...}{Any acceptable WebAPI parameters}

\item{token}{App ID}
}
\description{
Retrieve list of national wide BPS domain.
}
\examples{
\dontrun{
bps_get_domain(token = "your_secret_token")
}

}
