% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_get_subject.R
\name{bps_get_subject}
\alias{bps_get_subject}
\title{Get list of subject data}
\usage{
bps_get_subject(domain_id = "0000", ..., token)
}
\arguments{
\item{domain_id}{Domain id for single BPS website}

\item{...}{Any acceptable WebAPI parameters for subject data}

\item{token}{App ID}
}
\description{
Retrieve list of available subject data
}
\examples{
\dontrun{
bps_get_subject(token = "your_secret_token")
}

}
