% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_list_domain.R
\name{bps_list_domain}
\alias{bps_list_domain}
\title{Get list of domain}
\usage{
bps_list_domain(type = "all", ..., token)
}
\arguments{
\item{type}{Type to show domain}

\item{...}{Any acceptable WebAPI parameters: prov}

\item{token}{App ID}
}
\description{
Retrieve list of nation wide BPS domain.
}
\examples{
\dontrun{
bps_list_domain() # return all domains
bps_list_domain(type = "prov")
bps_list_domain(type = "kabbyprov", prov = 3200)
}

}
