% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_list_table.R
\name{bps_list_table}
\alias{bps_list_table}
\title{Get list of table}
\usage{
bps_list_table(keyword, model = "static", var, domain_id = "0000", ..., token)
}
\arguments{
\item{keyword}{Search parameter}

\item{model}{Table model can be static or dynamic}

\item{var}{Variable ID selected to display dynamic table data}

\item{domain_id}{Domain id for single BPS website}

\item{...}{Any acceptable WebAPI for table: page, year, month, lang}

\item{token}{App ID}
}
\description{
Retrieve list of table based on model
}
\examples{
\dontrun{
bps_list_table(keyword = "penduduk")
bps_list_table("penduduk", domain_id = "1100")
}

}
