% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_dataframe}
\alias{build_dataframe}
\title{Dataframe builder}
\usage{
build_dataframe(listdata)
}
\arguments{
\item{listdata}{List data from JSON}
}
\description{
Construct tibble dataframe from list of JSON
}
