% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_path}
\alias{build_path}
\title{Path builder}
\usage{
build_path(param, value = NULL, replace = TRUE)
}
\arguments{
\item{param}{Query parameter name}

\item{value}{Query parameter value. Used if replace = FALSE}

\item{replace}{Replacable value, default to TRUE}
}
\description{
Construct path for query
}
