% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_query}
\alias{build_query}
\title{Query builder}
\usage{
build_query(service, ..., token = NULL)
}
\arguments{
\item{service}{Service being used}

\item{...}{Any acceptable WebAPI parameters}

\item{token}{App ID}
}
\description{
Construct query for request
}
