% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_more_pages}
\alias{get_more_pages}
\title{Retrieve more pages}
\usage{
get_more_pages(query, pages, token = get_token())
}
\arguments{
\item{query}{Full query without token}

\item{pages}{Number of pages}

\item{token}{App ID}
}
\description{
Add token to the end of query
}
