% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{message_resp_error}
\alias{message_resp_error}
\title{Message to request failure}
\usage{
message_resp_error(status)
}
\arguments{
\item{status}{Response status}
}
\description{
Return information about request failure
}
