% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{message_snippet_tbl}
\alias{message_snippet_tbl}
\title{Snippet message for list result}
\usage{
message_snippet_tbl(params, meta)
}
\arguments{
\item{params}{string to key-value pair of arguments}

\item{meta}{metadata from response}
}
\description{
Return information about list search result
}
