% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_get_domain.R
\name{bps_get_domain}
\alias{bps_get_domain}
\title{Get list of domain}
\usage{
bps_get_domain(token)
}
\arguments{
\item{token}{App ID. Please register to https://webapi.bps.go.id/developer/register}
}
\description{
Retrieve list of national wide BPS domain.
}
\examples{
\dontrun{
bps_get_domain(token = Sys.getenv("BPS_APP_ID"))
}

}
